Inverted script for axe, so script on axe instead of trees, all trees choppable in mainscene, movement speed upgraded.
