+++ Assets/Materials/Explosion03_Mobile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d4d1e5bfd954e6543839701d63b2e73a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
