+++ Assets/Polygonal Particles.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 565d87717935e414fbc0d1cc5fea88cd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
