+++ Assets/Scripts/AxeInteraction.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 907f058f872a9d7409229951e36941c1
+guid: 7ec26ab6a89e9544f8a587492c786dab
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
