+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,10 +5,10 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/MainScene.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/trees-jens.unity
     guid: ba2cb49f2d7597744a3b4ca019695be3
   m_configObjects:
