+++ ProjectSettings/ProjectSettings.asset
@@ -137,10 +137,7 @@ PlayerSettings:
   vulkanEnableCommandBufferRecycling: 1
   loadStoreDebugModeEnabled: 0
   bundleVersion: 0.1
-  preloadedAssets:
-  - {fileID: -1143651508906075210, guid: ae4d11341c9caa7438317ca30eb005aa, type: 2}
-  - {fileID: 2793661551461622418, guid: 0f6504411038bf64faec2ecb431f5f05, type: 2}
-  - {fileID: 11400000, guid: 9a4f39558f07965449759dd4e6a73688, type: 2}
+  preloadedAssets: []
   metroInputSource: 0
   wsaTransparentSwapchain: 0
   m_HolographicPauseOnTrackingLoss: 1
